adds controllers seperated
